scripts/requirements.txt
pandas
pyarrow
SQLAlchemy
psycopg2-binary
python-dotenv